+++ CHANGELOG.rst
@@ -1,6 +1,17 @@
 Release History
 ===============
 
+dev
+---
+
+**API Changes (Backward Incompatible)**
+
+-
+
+**Bugfixes**
+
+-
+
 6.1.0 (2025-01-22)
 ------------------